+++ drivers/07_filter.cpp
@@ -9,7 +9,7 @@ using simd_examples::benchmark::ImageOptions;
 using simd_examples::benchmark::ParseResult;
 
 void write_csv(std::ostream& output, const ImageOptions& options,
-               double time, float result, float difference) {
+               double time, double result, float difference) {
     const std::size_t pixels =
         static_cast<std::size_t>(options.width) * options.height;
     output << "exercise,kernel,implementation,size,repetitions,time_ms,result,max_abs_difference\n";
@@ -50,7 +50,7 @@ int main(int argc, char** argv) {
             simd_examples::benchmark::implementation::blur_horizontal(
                 input_view, output_view);
         }, options.repetitions);
-    const float result = simd_examples::benchmark::checksum(
+    const double result = simd_examples::benchmark::checksum(
         output.begin(), output.end());
     const float difference = simd_examples::benchmark::max_abs_difference(
         output.data(), expected.data(), pixels);
